refactor: move functions to toolkits directory
