+++ camel/toolkits/functions/open_api_specs/coursera/__init__.py
