+++ camel/toolkits/functions/open_api_specs/coursera/openapi.yaml
