+++ camel/toolkits/functions/open_api_specs/create_qr_code/__init__.py
