+++ camel/toolkits/functions/open_api_specs/klarna/openapi.yaml
