+++ camel/toolkits/functions/open_api_specs/nasa_apod/openapi.yaml
