+++ camel/toolkits/functions/open_api_specs/outschool/ai-plugin.json
