+++ camel/toolkits/functions/open_api_specs/outschool/openapi.yaml
