+++ camel/toolkits/functions/open_api_specs/outschool/paths/__init__.py
