+++ camel/toolkits/functions/open_api_specs/security_config.py
