+++ camel/toolkits/functions/open_api_specs/speak/__init__.py
