+++ camel/toolkits/functions/open_api_specs/speak/openapi.yaml
