+++ camel/toolkits/functions/open_api_specs/web_scraper/ai-plugin.json
