+++ camel/toolkits/functions/openai_function.py
