+++ camel/toolkits/functions/slack_functions.py
@@ -24,7 +24,7 @@ if TYPE_CHECKING:
 
     from slack_sdk import WebClient
 
-from camel.functions import OpenAIFunction
+from camel.toolkits.functions import OpenAIFunction
 
 logger = logging.getLogger(__name__)
