+++ examples/function_call/github_examples.py
@@ -17,10 +17,10 @@ from colorama import Fore
 
 from camel.agents import ChatAgent
 from camel.configs import ChatGPTConfig
-from camel.functions import OpenAIFunction
 from camel.messages import BaseMessage
 from camel.models import ModelFactory
 from camel.toolkits import GithubToolkit
+from camel.toolkits.functions import OpenAIFunction
 from camel.types import ModelPlatformType, ModelType
 from camel.utils import print_text_animated
