+++ examples/function_call/openapi_function.py
@@ -13,8 +13,8 @@
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 from camel.agents import ChatAgent
 from camel.configs.openai_config import ChatGPTConfig
-from camel.functions import OPENAPI_FUNCS
 from camel.messages import BaseMessage
+from camel.toolkits.functions import OPENAPI_FUNCS
 
 # Define system message
 sys_msg = BaseMessage.make_assistant_message(
